+++ balrogscript/task.py
@@ -5,13 +5,16 @@ import os
 import re
 import sys
 
+from balrogscript.constants import VALID_ACTIONS
+
 log = logging.getLogger(__name__)
 
 
-def validate_task_schema(script_config, task_definition):
+def validate_task_schema(script_config, task_definition, action):
     """Perform a schema validation check against taks definition"""
     schema_path = os.path.join(
-        os.path.dirname(os.path.dirname(__file__)), script_config['schema_file']
+        os.path.dirname(os.path.dirname(__file__)),
+        script_config['schema_files'][action]
     )
     with open(schema_path) as fh:
         schema = json.load(fh)
@@ -32,11 +35,24 @@ def get_task(script_config):
     return task_definition
 
 
-def get_task_channel(task, script_config):
-    """Extract task channel from scopes"""
-    # TODO to be implemented once balrogscript needs to handle rules munging
-    # too
-    raise NotImplementedError("This method has yet to be implemented")
+def get_task_action(task, script_config):
+    """Extract task server from scopes"""
+    actions = [
+        s.split(':')[-1] for s in task["scopes"] if
+        s.startswith("project:releng:balrog:action:")
+    ]
+    if actions:
+        log.info("actions: %s", actions)
+        if len(actions) != 1:
+            raise ValueError("Only one action can be used")
+        action = actions[0]
+    else:
+        action = 'submit-locale'
+
+    if action not in VALID_ACTIONS:
+        raise ValueError("Invalid action scope")
+
+    return action
 
 
 def get_task_server(task, script_config):