+++ balrogscript/test/test_task.py
@@ -2,18 +2,17 @@
 import pytest
 
 from balrogscript.test import config, nightly_config, release_config
-from balrogscript.task import (get_task, get_task_channel, get_task_server,
-                               get_upstream_artifacts, get_manifest)
+from balrogscript.task import (
+    get_manifest,
+    get_task,
+    get_task_action,
+    get_task_server,
+    get_upstream_artifacts,
+)
 
 assert nightly_config  # silence pyflakes
 assert release_config  # silece pyflakes
 assert config  # silence pyflakes
-
-
-def test_get_task_channel(nightly_config):
-    task = get_task(nightly_config)
-    with pytest.raises(NotImplementedError):
-        get_task_channel(task, nightly_config)
 
 
 @pytest.mark.parametrize("scopes,expected,raises", ((
@@ -97,3 +96,36 @@ def test_release_get_manifest(release_config):
         get_manifest(release_config, upstream_artifacts)
         assert e.type == SystemExit
         assert e.value.code == 3
+
+
+@pytest.mark.parametrize("task,expected,raises", ((
+    {"scopes": ["project:releng:balrog:action:submit-locale"]},
+    "submit-locale",
+    False
+), (
+    {"scopes": ["project:releng:balrog:action:submit-toplevel"]},
+    "submit-toplevel",
+    False
+), (
+    {"scopes": ["project:releng:balrog:action:schedule"]},
+    "schedule",
+    False
+), (
+    {"scopes": ["project:releng:balrog:action:schedule", "project:releng:balrog:action:submit-locale"]},
+    None,
+    True
+), (
+    {"scopes": ["project:releng:balrog:action:illegal"]},
+    None,
+    True
+), (
+    {"scopes": []},
+    'submit-locale',
+    False
+)))
+def test_get_task_action(task, expected, raises):
+    if raises:
+        with pytest.raises(ValueError):
+            get_task_action(task, {})
+    else:
+        assert get_task_action(task, {}) == expected